pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hawk-hooks"
version = "0.5.0"
description = "A modular Claude Code hooks manager"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Peter Kronström", email = "peter.kronstrom@gmail.com" }
]
keywords = ["claude", "hooks", "cli", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "questionary>=2.0.0",
    "rich>=13.0.0",
    "readchar>=4.0.0",
    "pyyaml>=6.0",
    "simple-term-menu>=1.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
hawk-hooks = "hawk_hooks.cli:main"
hawk = "hawk_hooks.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/hawk_hooks", "src/rich_menu"]

[tool.hatch.build.targets.wheel.force-include]
"examples" = "hawk_hooks/examples"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --tb=short"
